Split stub classes into separate files
greppability is king
