+++ lib/miniproxy/stub/request.rb
@@ -20,26 +20,5 @@ module MiniProxy
         http_request.request_method == @method && request_uri.match?(@url)
       end
     end
-
-    # MiniProxy stub response, so stubbed requests can return a custom response
-    #
-    class Response
-      def initialize(headers:, body:)
-        @body = body
-        @headers = headers
-      end
-
-      def body
-        @body || ""
-      end
-
-      def code
-        200
-      end
-
-      def headers
-        @headers || { "Content-Type" => "text/html" }
-      end
-    end
   end
 end
