+++ spec/lib/miniproxy/stub/request_spec.rb
@@ -1,0 +1,1 @@
+require "miniproxy/stub/request"
