+++ spec/lib/miniproxy/stub/response_spec.rb
@@ -1,0 +1,1 @@
+require "miniproxy/stub/response"
