--- spec/lib/miniproxy/stub_spec.rb
@@ -1,1 +1,0 @@
-require "miniproxy/stub"
